% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cbsa_coord.R
\docType{data}
\name{cbsa_coord}
\alias{cbsa_coord}
\title{CBSA name and coordinates}
\format{
A data.table with 938 rows and 4 variable:
\describe{
  \item{msa_fips}{CBSA FIPS}
  \item{msa}{CBSA name}
  \item{lon}{longitude}
  \item{lat}{latitude}
}
}
\usage{
data("cbsa_coord")
}
\description{
Extract from 2019 tiger shape files
}
\keyword{datasets}
