% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_state_names.R
\name{convert_state_names}
\alias{convert_state_names}
\title{Convert state names among full, lowercase, and abbreviation}
\usage{
convert_state_names(states, to)
}
\arguments{
\item{states}{string, vector of state names to be converted.}

\item{to}{string, format after convertion, take values of "full",
"lowercase", and "abbr".}
}
\value{
a vector of converted state names
}
\description{
Also include other areas such as Puerto Rico and Guam, and even United states
for dealing with census data. Examples of state names:
full      --  "District of Columbia", "Kentucky";
lowercase --  "district of columbia", "kentucky";
abbr      --  "DC", "KY".
}
\examples{
convert_state_names(c("Alabama",  "California","Connecticut"), "abbr")

# wrong names return NA and a warning
convert_state_names(c("DC", "AA","MA"), "full")


}
