% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table.R
\name{download_table}
\alias{download_table}
\title{Download a online table}
\usage{
download_table(url, xpath, save_as)
}
\arguments{
\item{url}{url of the webpage containing the table}

\item{xpath}{xpath of the table}

\item{save_as}{save table as "save_as.csv" file}
}
\value{
a data frame or a saved csv file
}
\description{
Download a online table
}
\details{
Inspect the webpage, highlight css code for table, right click then
copy --> copy XPath, and paste inside SINGLE quote ' ' as xpath contains
double quotes.

Package xml2 is required for rvest. Need to import it explicitly in the package.
}
\examples{
table_df <-  download_table('https://en.wikipedia.org/wiki/Judgment_of_Paris_(wine)',
                           '//*[@id="mw-content-text"]/div/table[5]',
                           "wine.csv")
}
