% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sf_columns.R
\name{extract_sf_columns}
\alias{extract_sf_columns}
\title{Extract columns from a sf object and return a dataframe}
\usage{
extract_sf_columns(dat_sf, cols)
}
\arguments{
\item{dat_sf}{a sf data frame}

\item{cols}{string vector of selected columns}
}
\value{
a non-sf data frame
}
\description{
Extract columns from a sf object and return a dataframe
}
