% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakeTextFile.R
\name{fakeTextFile}
\alias{fakeTextFile}
\title{Create a dummy text file in a directory.}
\usage{
fakeTextFile(
  file_extension = ".R",
  dir_path = NULL,
  file_path = NULL,
  txt = NULL,
  n_lines = 1
)
}
\arguments{
\item{file_extension}{Character value. File extension, default value is '.R'}

\item{dir_path}{A character vector with the path to the folder where to save the text file.
Default value is NULL and uses \code{\link[base:tempfile]{base::tempdir}}.}

\item{file_path}{A character vector with the path, file name and extension of the created random file.}

\item{txt}{An optional string (one line) or list of strings (multiple lines)
to save in the text file. Randomly generated by default}

\item{n_lines}{Positive none zero integer value. If `txt` is not provided, `n_lines` is the number of lines of text in the file.}
}
\value{
A list with two elements:
 - tmp_path: the path to the fake text file
 - txt: the dummy text string or list of strings
}
\description{
Creates a dummy string of text, saves a text file in \code{\link[base:tempfile]{base::tempdir}}.
}
\details{
If not working inside an R project, dir_path has to be complete from root.
   To create an empty file, use `txt = ""`.
}
\examples{
\dontrun{fakeTextFile()}
}
\seealso{
For more flexibility in random string generation, see \code{NCmisc::fakeLines}.
   For more automated paragraph generation, see \code{stringi::stri_rand_lipsum}.
}
\author{
Alban Sagouis
}
