% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation_in_scripts.r
\name{gsubInOneFolder}
\alias{gsubInOneFolder}
\title{Replace text in all scripts from a folder}
\usage{
gsubInOneFolder(fullPath, pattern, replacement, recursive = FALSE)
}
\arguments{
\item{fullPath}{A character vector with the path to the folder countaining
the scripts to modify.}

\item{pattern}{passed to /code{gsub}.}

\item{replacement}{passed to /code{gsub}.}

\item{recursive}{logical. Passed to list.files. Should the listing recurse into directories?}
}
\description{
Open a file with a .r or .R extension and runs a gsub.
}
\details{
If not working inside an Rproject, path has to be complete from root.
}
\examples{
fullPath <- "C:/Users/as80fywe/idiv/my r packages/myToolbox/tests/testthat/objects/"
fullPath <- "./tests/testthat/objects/"
gsubInOneScript(fullPath, pattern="random_2020", replacement = "newID_2021")

}
\author{
Alban Sagouis
}
