% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_version_numbers.r
\name{incVer}
\alias{incVer}
\title{Increment version number of a package}
\usage{
incVer(pkg, folder = ".", increase = "dev")
}
\arguments{
\item{pkg}{a character vector containing complete}

\item{folder}{description}

\item{increase}{description}
}
\description{
Reads DESCRIPTION from installed package 'pkg' and makes a new one in the specified
'folder'. Three options for 'increase' are "minor", "patch" and "dev" respectively
incrementing the second, third and fourth elements of version 0.1.0.9012.
}
\details{
Also updates the date to system date.

description. Originally posted by alko989 here https://stackoverflow.com/q/24209336. Hadley Wickham discusses when and how increment version numbers in his book http://r-pkgs.had.co.nz/release.html#release-version
}
\author{
alko989, Alban Sagouis
}
