% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_base_US.R
\name{map_base_usa}
\alias{map_base_usa}
\title{base ggplot2 map of all states}
\usage{
map_base_usa(textsize = 5, textcolor = "grey")
}
\arguments{
\item{textsize}{integer, text size}

\item{textcolor}{string, text color}
}
\value{
a ggplot2 objecct
}
\description{
The map is based on the data in package albersusa
}
