% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_base_US.R
\name{map_basse_US}
\alias{map_basse_US}
\title{base ggplot2 map of all states}
\usage{
map_basse_US(textsize = 5, textcolor = "grey")
}
\value{
a ggplot2 objecct
}
\description{
The map is based on the data in package albersusa
}
