% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_move_al_hi.R
\name{map_move_ak_hi}
\alias{map_move_ak_hi}
\title{Move the locations in Alaska and Hawaii}
\usage{
map_move_ak_hi(df, x = "lon", y = "lat")
}
\arguments{
\item{df}{a data.frame containing coordinates}

\item{x}{specify longitude column}

\item{y}{specify latitude column}
}
\value{
The same data frame with altered coordimates in AK and HI
}
\description{
Move (lon, lat) in AK and HI to new locations match the state
map provided by the albersusa package (github)
}
\details{
This function will identify those coordinates belongs to AK or HI
and moves them to the US map provided by the albersusa package. Coordinates
belong to other states will not moved.
}
\examples{
counties <- totalcensus::read_acs1year(
  2019, c("AK", "HI", "RI"), summary_level = "county"
)

counties_moved <- map_move_ak_hi(counties)

}
