% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisePath.r
\name{normalisePath}
\alias{normalisePath}
\title{Replace backslashes into slashes in paths}
\usage{
normalisePath(fullPath)
}
\arguments{
\item{fullPath}{A character vector with the path to transform.}
}
\value{
Character string: transformed path.
}
\description{
Replace backslashes into slashes in paths
}
