% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfse.lines.r
\name{readlines}
\alias{readlines}
\title{readlines}
\usage{
readlines(x, ...)
}
\arguments{
\item{x}{Input}

\item{...}{Other args passed to \code{readLines}.}
}
\value{
Output
}
\description{
Read lines of file
}
\details{
Simple wrapper around \code{\link[base]{readLines}} that automates opening and
  closing of connection file.
}
\author{
Michael Wayne Kearney
}
