% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfse.search_files.r
\name{search_files}
\alias{search_files}
\title{search_files}
\usage{
search_files(
  x,
  path = ".",
  recursive = TRUE,
  all.files = FALSE,
  textOnly = TRUE
)
}
\arguments{
\item{x}{Pattern.}

\item{path}{Path on which to restrict search. Defaults to current working
directory.}

\item{recursive}{logical}

\item{all.files}{logical. Default \code{FALSE} excludes dot files}

\item{textOnly}{logical. Default \code{TRUE}. Looks inside .r .R
.Rmd .md .html .txt .cpp .h files.}
}
\value{
Output from terminal - file name, line number, and preview of
  matching text
}
\description{
Returns matching files and line numbers of given string pattern.
}
