% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfse.search_files.r
\name{search_these_files}
\alias{search_these_files}
\title{Search these files}
\usage{
search_these_files(x, f, ignore.case = TRUE)
}
\arguments{
\item{x}{Regex pattern}

\item{f}{Vector of files}

\item{ignore.case}{logical. FALSE for case-sensitive matches. Default is TRUE.}
}
\value{
Prints matches
}
\description{
Look for text in a group of files
}
\author{
Michael Wayne Kearney
}
