% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_state_center_coord.R
\docType{data}
\name{state_center}
\alias{state_center}
\title{State center coordinates}
\format{
A data.table with 51 rows and three variable:
\describe{
  \item{state}{postal code of state such as "CA" and "TX"}
  \item{lon}{longitude}
  \item{lat}{latitude}
}
}
\usage{
data("state_center")
}
\description{
It is close to true geographic center of a state
}
\keyword{datasets}
