% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_state_center_coord_for_label.R
\docType{data}
\name{state_center_lab}
\alias{state_center_lab}
\title{State center coordinates for label}
\format{
A data.table with 51 rows and three variable:
\describe{
  \item{state}{postal code of state such as "CA" and "TX"}
  \item{lon}{longitude}
  \item{lat}{latitude}
}
}
\usage{
data("state_center_lab")
}
\description{
This data is slightly different from the trun geographic state center as some
small state at east coast have been offset for better labeling.
}
\keyword{datasets}
