% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_state_center_moved.R
\docType{data}
\name{state_center_moved}
\alias{state_center_moved}
\title{State center coordinates moved according to albersusa package}
\format{
A data.table with 51 rows and three variable:
\describe{
  \item{state}{postal code of state such as "CA" and "TX"}
  \item{lon}{longitude, directly from ACS data}
  \item{lat}{latitude, directly from ACS data}
  \item{lon1}{longitude, shift some states for text label}
  \item{lat1}{latitude, shift some states for text label}
  \item{lon2}{longitude, shift some states for line label}
  \item{lat2}{latitude, shift some states for line label}
}
}
\usage{
data("state_center_moved")
}
\description{
This data is slightly different from the trun geographic state center as some
small state at east coast have been offset for better labeling.
}
\keyword{datasets}
