% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeout.R
\name{timeout}
\alias{timeout}
\title{Stop the execution of an expression if the elapsed time exceed a limit}
\usage{
timeout(expr, timeout)
}
\arguments{
\item{expr}{R expression such as sum(1:5) and muli-line R code inside {}.}

\item{timeout}{numeric, time limit in second}
}
\value{
return of expr
}
\description{
Stop the execution of an expression if the elapsed time exceed a limit
}
\examples{
timeout(
  {
    for (i in 1:1e8) {
      print(i)
    }
  },
  timeout = 0.1
)

}
