% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeAll.r
\name{executeAll}
\alias{executeAll}
\title{Execute all scripts inside a folder}
\usage{
executeAll(fullPath = NULL, echo = TRUE, local = TRUE)
}
\arguments{
\item{fullPath}{A character vector with the path to the folder containing the scripts to
execute}

\item{echo}{logical. Passed to \code{\link[base]{source}}.}

\item{local}{logical. Passed to \code{\link[base]{source}}.}
}
\description{
Source all files with a .r or .R extension in the specified folder.
}
\details{
If not working inside an Rproject, path has to be complete from root.
}
\author{
Alban Sagouis
}
