% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeAll.r
\name{executeOne}
\alias{executeOne}
\title{Execute one script in the background without loading anything in the user's current environment}
\usage{
executeOne(fullPath = NULL, echo = FALSE, local = TRUE)
}
\arguments{
\item{fullPath}{A character vector with the path to the scripts to execute}

\item{echo}{logical. Passed to \code{\link[base]{source}}.}

\item{local}{logical. Passed to \code{\link[base]{source}}.}
}
\description{
Source the given file with a .r or .R extension.
}
\details{
If not working inside an Rproject, path has to be complete from root.
}
\author{
Alban Sagouis
}
