% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation_in_scripts.r
\name{gsubInOneFolder}
\alias{gsubInOneFolder}
\title{Replace text in all scripts from a folder}
\usage{
gsubInOneFolder(
  fullPath,
  pattern,
  replacement,
  recursive = FALSE,
  newPath = NULL
)
}
\arguments{
\item{fullPath}{A character vector with the path to the folder countaining
the scripts to modify.}

\item{recursive}{logical. Passed to \code{\link[base]{list.files}}. Should the listing recurse into directories?}

\item{newPath}{description}
}
\description{
Open a file with a .r or .R extension and runs a gsub.
}
\details{
If not working inside an Rproject, path has to be complete from root.
}
\author{
Alban Sagouis
}
