% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation_in_scripts.r
\name{gsubInOneFolder}
\alias{gsubInOneFolder}
\title{Replace text in all scripts from a folder}
\usage{
gsubInOneFolder(
  fullPath,
  pattern,
  replacement,
  recursive = FALSE,
  newPath = NULL,
  userCheck = TRUE
)
}
\arguments{
\item{fullPath}{A character vector with the path to the folder containing
the scripts to modify.}

\item{pattern}{passed to \code{\link[base:grep]{base::gsub}}.}

\item{replacement}{passed to \code{\link[base:grep]{base::gsub}}.}

\item{recursive}{logical. Passed to \code{\link[base:list.files]{base::list.files}}. Should the listing recurse into directories?}

\item{newPath}{description}

\item{userCheck}{Logical, default = TRUE. Matches are shown to check the pattern `accuracy`
before replacement.}
}
\value{
A data.frame with the number of matches in each file.
}
\description{
Opens all files with a .r or .R extension in the directory and runs a gsub.
}
\details{
If not working inside an Rproject, path has to be complete from root.
}
\examples{
\dontrun{gsubInOneFolder('./R', pattern = 'sum', replacement = 'sumNA')}
\dontrun{gsubInOneFolder('.',  pattern = 'read.csv', replacement = 'data.table::fread')}

}
\author{
Alban Sagouis
}
