% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_manipulation_in_scripts.r
\name{gsubInOneScript}
\alias{gsubInOneScript}
\title{Replace text in one script}
\usage{
gsubInOneScript(
  fullPath,
  pattern,
  replacement,
  newPath = NULL,
  userCheck = TRUE,
  ignore.case = FALSE
)
}
\arguments{
\item{fullPath}{A character vector with the path to the script to modify.}

\item{pattern}{passed to \code{\link[base:grep]{base::gsub}}.}

\item{replacement}{passed to \code{\link[base:grep]{base::gsub}}.}

\item{newPath}{Character vector. Used if a copy of the script should be created
following this path.}

\item{userCheck}{Logical, default = TRUE. Matches are shown to check the pattern `accuracy`
before replacement.}

\item{ignore.case}{logical. FALSE for case-sensitive matches. Default is TRUE.}
}
\value{
The number of matches in the file.
}
\description{
Open a file with a .r or .R extension and runs a gsub. No copy made except if newPath is
specified, use at your own risks.
}
\details{
If not working inside an R project, path has to be complete from root.
}
\examples{
\dontrun{gsubInOneScript('DESCRIPTION','foobar@gmail.com','foobar@protonmail.com')}
\dontrun{gsubInOneScript('master.r','ifelse','data.table::fifelse')}
}
\seealso{
To search for text in any text file, see \code{search_files}.
}
\author{
Alban Sagouis
}
