% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_version_numbers.r
\name{incVer}
\alias{incVer}
\title{Increment version number of a package}
\usage{
incVer(pkg, folder = ".", increase = "dev")
}
\arguments{
\item{pkg}{A character vector withthe name of the package to update.}

\item{folder}{Folder in which DESCRIPTION is saved. Should be left as '.'.}

\item{increase}{Character string, either major, minor, patch or dev.}
}
\description{
Reads DESCRIPTION from installed package 'pkg' and makes a new one in the specified
'folder'. Four options for 'increase' are "major", minor", "patch" and "dev"
respectively incrementing the first to fourth element of version 0.1.0.9012.
Also changes the date to system date.
}
\details{
Function originally posted by alko989 here https://stackoverflow.com/q/24209336. Hadley Wickham discusses when and how increment version numbers in his book http://r-pkgs.had.co.nz/release.html#release-version
}
\author{
alko989, Alban Sagouis
}
